# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/role-permissions": {
      "get": {
        "operationId": "AccessControlController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "access-control"
        ]
      }
    },
    "/api/v1/role-permissions/{role}": {
      "patch": {
        "operationId": "AccessControlController_update",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "access-control"
        ]
      }
    },
    "/api/v1/dictionaries/{type}": {
      "get": {
        "operationId": "DictionariesController_list",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "sizes",
                "managers",
                "priorities",
                "statuses",
                "weights"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      },
      "post": {
        "operationId": "DictionariesController_create",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "sizes",
                "managers",
                "priorities",
                "statuses",
                "weights"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/{type}/{id}": {
      "patch": {
        "operationId": "DictionariesController_update",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "sizes",
                "managers",
                "priorities",
                "statuses",
                "weights"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictionaryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      },
      "delete": {
        "operationId": "DictionariesController_remove",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "departments",
                "sizes",
                "managers",
                "priorities",
                "statuses",
                "weights"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/weights/{id}/apply-to-open-cards": {
      "post": {
        "operationId": "DictionariesController_applyWeight",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/dictionaries/sizes/recalculate-open-cards": {
      "post": {
        "operationId": "DictionariesController_recalculateSizes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "dictionaries"
        ]
      }
    },
    "/api/v1/custom-fields": {
      "get": {
        "operationId": "CustomFieldsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      },
      "post": {
        "operationId": "CustomFieldsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessDto"
                }
              }
            }
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/api/v1/custom-fields/{id}": {
      "patch": {
        "operationId": "CustomFieldsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      },
      "delete": {
        "operationId": "CustomFieldsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "custom-fields"
        ]
      }
    },
...
    "/api/v1/backups/export": {
      "get": {
        "operationId": "DataManagementController_export",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "data-management"
        ]
      }
    },
    "/api/v1/backups/validate": {
      "post": {
        "operationId": "DataManagementController_validate",
        "parameters": [
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "data-management"
        ]
      }
    },
    "/api/v1/backups/import": {
      "post": {
        "operationId": "DataManagementController_import",
        "parameters": [
          {
            "name": "mode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupImportRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "data-management"
        ]
      }
    },
...
  },
  "info": {
    "title": "PMO Hub API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "new_password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "current_password",
          "new_password"
        ]
      },
      "ApiSuccessDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true,
            "enum": [
              true
            ]
          },
          "message": {
            "type": "string",
            "example": "Операцію виконано"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "success"
        ]
      },
...
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "canCreateEditProjects": {
            "type": "boolean"
          },
          "canDeleteProjects": {
            "type": "boolean"
          },
          "canAccessAdmin": {
            "type": "boolean"
          },
          "isReadOnly": {
            "type": "boolean"
          },
          "canEditArchive": {
            "type": "boolean"
          }
        }
      },
      "DictionaryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "capacity_limit_points": {
            "type": "number",
            "minimum": 0
          },
          "department_id": {
            "type": "string",
            "format": "uuid"
          },
          "color": {
            "type": "string",
            "pattern": "^#?([0-9A-F]{3}|[0-9A-F]{4}|[0-9A-F]{6}|[0-9A-F]{8})$"
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "min_score": {
            "type": "number",
            "minimum": 0
          },
          "max_score": {
            "type": "number",
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          },
          "code": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
      "CustomFieldDto": {
        "type": "object",
        "properties": {
          "entityType": {
            "type": "string",
            "enum": [
              "project",
              "task"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "NUMBER",
              "SELECT",
              "CHECKBOX",
              "RICHTEXT"
            ]
          },
          "isRequired": {
            "type": "boolean"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "showInTable": {
            "type": "boolean"
          },
          "showInCards": {
            "type": "boolean"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "entityType",
          "name",
          "type",
          "isRequired"
        ]
      },
      "BackupImportRequestDto": {
        "type": "object",
        "properties": {
          "backup": {
            "type": "object",
            "additionalProperties": true
          },
          "validation_token": {
            "type": "string"
          }
        },
        "required": [
          "backup",
          "validation_token"
        ]
      }
    }
  }
}